Stop typing __all__. Rename .middlewares to .middleware and .auths to .auth
