+++ neoclient/sentinels.py
@@ -1,5 +1,3 @@
-from typing import Sequence
-
 from pydantic.fields import Required, Undefined
 
-__all__: Sequence[str] = ("Required", "Undefined")
+__all__ = ("Required", "Undefined")
