+++ tests/test_decorators.py
@@ -6,9 +6,9 @@ from httpx import URL, Cookies, Headers, QueryParams
 from pytest import fixture
 
 from neoclient import converters, decorators, get
-from neoclient.auths import Auth, BasicAuth
+from neoclient.auth import Auth, BasicAuth
 from neoclient.defaults import DEFAULT_FOLLOW_REDIRECTS
-from neoclient.middlewares import AuthMiddleware
+from neoclient.middleware import AuthMiddleware
 from neoclient.models import ClientOptions, RequestOpts, Request, Response, State
 from neoclient.operation import get_operation
 from neoclient.services import Service
